Refactored unit tests to use new layout.
