+++ pallets/stateful-storage/src/benchmarking.rs
@@ -307,6 +307,6 @@ benchmarks! {
 	}
 
 	impl_benchmark_test_suite!(StatefulStoragePallet,
-		crate::mock::new_test_ext_keystore(),
-		crate::mock::Test);
+		crate::tests::mock::new_test_ext_keystore(),
+		crate::tests::mock::Test);
 }
