+++ pallets/stateful-storage/src/tests/mod.rs
@@ -1,0 +1,5 @@
+pub mod mock;
+#[cfg(any(feature = "runtime-benchmarks", test))]
+mod test_common;
+
+mod other_tests;
